Added new test, splitted objective
