+++ monolith/views/__init__.py
@@ -4,5 +4,6 @@ from .users import users
 from .strava import strava
 from .comparisons import comparisons
 from .runs import runs
+from .objective import objective
 
-blueprints = [home, auth, users, strava, comparisons, runs]
+blueprints = [home, auth, users, strava, comparisons, runs, objective]
