+++ monolith/views/home.py
@@ -71,25 +71,3 @@ def index():
         progress = m2km(progress),
         percentage = percentage
     )
-
-
-@home.route('/objective', methods=['GET', 'POST'])
-def set_objective():
-    form = ObjectiveForm()
-
-    if form.validate_on_submit():
-        objective_distance = km2m(form.data['distance'])
-        q = db.session.query(User).filter(User.email == current_user.email)
-        user = q.first()
-
-        existing_objective = db.session.query(Objectives).filter(Objectives.user == user).first()
-        if existing_objective is None:
-            _setObjective(user, objective_distance)
-
-        else:
-            existing_objective.set_distance(objective_distance)
-
-        db.session.commit()
-        return redirect("/")
-
-    return render_template('set_objective.html', form = form)
